requirements.txt
aiohttp==0.22.5
chardet==2.3.0
Jinja2==2.8
MarkupSafe==0.23
multidict==1.2.2
pytaglib==1.3.0
